dpa_offload: Add support for header manipulation on DPA Classifier next_table action
The DPA Classifier API did not allow the user to add header
manipulation on an entry whose action was sending the frames to
a new classification table. This change adds this support
to the DPA Classifier by adding an "hmd" (header manipulation
descriptor) attribute to the "dpa_cls_tbl_next_table_desc"
parameters structure.

Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Change-Id: I126cdfad8e76a8bec4453c8f7b506ad3380349e2
Reviewed-on: http://git.am.freescale.net:8181/3106
Reviewed-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>
Tested-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>