Set SDQCR to ensure frame queues can reach the retired state
In some situations frame queues may still be on a work queue when
a retirement command is sent during cleanup. In order to reach the
retired state these frame queues must be scheduled. This patch sets
an appropriate SDQCR value to ensure that the Frame Queue is scheduled.

Since Frame Queues could be 'locked' to a particular portal if the
Hold Active feature is enabled the shutrown routine must service
all portals to ensure that the FQ reaches a retired state

Signed-off-by: Roy Pledge <Roy.Pledge@freescale.com>
Change-Id: I97f79d35c62f179f21b6bd11c23bc0150e54ff5f
Reviewed-on: http://git.am.freescale.net:8181/3339
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>
Tested-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>