dpa_offload: dpa_ipsec - Update create SA function to set hmd to invalid
DPA Classifier API was updated by adding a new attribute (hmd)
to the struct dpa_cls_tbl_next_table_desc.

DPA IPSec is using the classifier feature for linking two tables,
if inbound policy verification is required but does not require
the hmd option.

A minor update had to be made to set the hdm
to -1 (not valid) otherwise the 0 value would have been concidered
valid by classifier and crash since not hmd is required here.

Signed-off-by: Andrei Varvara <andrei.varvara@freescale.com>
Change-Id: I76eccbf6a60b7d563540de539be5144112c1567c
Reviewed-on: http://git.am.freescale.net:8181/3474
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>