dpaa_eth: Remove dpa_bp_count from percpu_priv structure
Instead just rely on the percpu field in dpa_bp structure.
The performance lost due to the extra indirection is compensated
by the smaller size of the percpu_priv structure.

This way we begin to eliminate the assumption that private interfaces
only have one default buffer pool, which may not remain true for long.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@freescale.com>
Change-Id: I0b2c199c3f8883fa7d5fa1a06d2940f8170fc959
Reviewed-on: http://git.am.freescale.net:8181/3655
Reviewed-by: Bucur Madalin-Cristian-B32716 <madalin.bucur@freescale.com>
Reviewed-by: Hamciuc Bogdan-BHAMCIU1 <bogdan.hamciuc@freescale.com>
Reviewed-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>
Tested-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>