dpaa_eth: Fix draining of buffer pools
When draining buffers in batches of eight, we must take into account
the fact that BMan will return an error code if there are less than
eight buffers left in the pool; when this happens, drain the remaining
buffers one by one.

And since we're at buffer pool cleanup, make sure allocated bpools get
freed in case of errors inside the probe function.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@freescale.com>
Change-Id: Id0ae7ed369ded0fff7359ae20b6bf5df64e77fa1
Reviewed-on: http://git.am.freescale.net:8181/4442
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Bucur Madalin-Cristian-B32716 <madalin.bucur@freescale.com>
Reviewed-by: Sovaiala Cristian-Constantin-B39531 <Cristian.Sovaiala@freescale.com>
Reviewed-by: Rivera Jose-B46482 <German.Rivera@freescale.com>