dpa_offload: Added Traffic Manager user-space counter implementation
Added support for Traffic Manager user-space DPA Stats counter in
driver. The Traffic Manager counters can be used both from user-space or
kernel-space depending on the user needs. If the counter is created from
user-space, all the retrieving mechanism is implemented in user-space
and same happens for kernel-space case. The counters is marked as a
user-space counter and the kernel will provide for this counter only the
offset where the user-space library to write the statistics of the counter.
    This patch corrects some compilation warnings caused by QMan CEETM
API changes.

Change-Id: I48c4df286ab78efe2a16d46c4fd376d153732944
Signed-off-by: Aurelian Zanoschi <Aurelian.Zanoschi@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/4196
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Chereji Marian-Cornel-R27762 <marian.chereji@freescale.com>
Reviewed-by: Rivera Jose-B46482 <German.Rivera@freescale.com>