fsl_dce: Add support for stateless compression/decompression
The fsl_dce device does compression and decompression of
various formats (deflate, zlib, gzip).
The flib directory contains the base layer object required
to access the dce device. The fsl_dce_flow object contains the
minimum set of interfaces required to send and receive frames
to/from the dce via QMan.

The fsl_dce_chunk object is an extension of the fsl_dce_flow object.
This object is used to compress or decompress single frames (chunks)
at a time.

Signed-off-by: Jeffrey Ladouceur <Jeffrey.Ladouceur@freescale.com>
Change-Id: Ia1f671b4610fb70d50353385a3d87dde0584741a
Reviewed-on: http://git.am.freescale.net:8181/4295
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Thorpe Geoff-R01361 <Geoff.Thorpe@freescale.com>
Reviewed-by: Rivera Jose-B46482 <German.Rivera@freescale.com>