dpaa_eth: Use one NAPI structure per portal
We currently use one NAPI instance per CPU which is enough when
having a single portal affined to each CPU.
In CPU Hotplug scenario, when CPUs become offline, portals affined
to them are migrated to the boot CPU therefore CPU0 will receive
interrupts not only from its affined portal but also from the
portal(s) that have migrated to it. In this case, a single NAPI
structure is not sufficient, therefore we introduce a list of
NAPI structures per CPU per portal such that when an interrupt
is received from a specific portal, we retrieve and schedule the
NAPI corresponding to that CPU and that portal.

Signed-off-by: Cristian Sovaiala <cristian.sovaiala@freescale.com>
Change-Id: I837e007a4256aba31bb1df9f893a20a1b0160a32
Reviewed-on: http://git.am.freescale.net:8181/5628
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Bucur Madalin-Cristian-B32716 <madalin.bucur@freescale.com>
Reviewed-by: Ruxandra Ioana Radulescu <ruxandra.radulescu@freescale.com>
Reviewed-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>
Reviewed-by: Rivera Jose-B46482 <German.Rivera@freescale.com>