dpaa_eth: Add pool channel in each affined portal's SDQCR
There might be CPUs which are not enabled at boottime, therefore
the pool channel is not currently set to the SDQCR of the affined
portals of those CPUs.
This means that the portals affined to the offline CPUs and which
are migrated to the boot CPU, will not issue interrupts to the
boot CPU in first instance and to the affined CPU (when this comes
online) in the second instance.

Therefore we set the pool channel to the SDQCR of each affined portal.

Signed-off-by: Cristian Sovaiala <cristian.sovaiala@freescale.com>
Change-Id: I63f5649e0c2169617dff991f08dd872a435578a8
Reviewed-on: http://git.am.freescale.net:8181/5629
Reviewed-by: Bucur Madalin-Cristian-B32716 <madalin.bucur@freescale.com>
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Ruxandra Ioana Radulescu <ruxandra.radulescu@freescale.com>
Reviewed-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>
Reviewed-by: Rivera Jose-B46482 <German.Rivera@freescale.com>