fmd: Fix problem in dynamic resource allocation algorithm
The patch that removed the forwarding optimized DPAA Ethernet driver
(and implicitly the Kconfig option CONFIG_FSL_DPAA_ETH_SG_SUPPORT and
all associated code) deleted the wrong branch of an #ifdef in the fmd
dynamic resource allocation algorithm.

As a result, the allocated FIFO size was not computed correctly,
causing resource depletion in case jumbo frames support is enabled.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@freescale.com>
Change-Id: Id052d76ef397c8865865cb7b185dc147fa5b154c
Reviewed-on: http://git.am.freescale.net:8181/6543
Reviewed-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Jose Rivera <German.Rivera@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/7675
Reviewed-by: Madalin-Cristian Bucur <madalin.bucur@freescale.com>
Tested-by: Madalin-Cristian Bucur <madalin.bucur@freescale.com>