mtd: m25p80: assign default read command
In the following commit (in -next):

    commit 8552b439aba7f32063755d23f79ca27b4d0a3115
    drivers: mtd: m25p80: convert "bool" read check into an enum

We converted the boolean 'fast_read' property to become an enum
'flash_read', but at the same time, we changed the conditional path so
that it doesn't choose a default value in some cases (technically, we
choose the correct default simply by virtue of devm_kzalloc(), which
zeroes this out to be a NORMAL read operation, but still...).

Fix this by setting a default for the 'else' clause.

Signed-off-by: Brian Norris <computersforpeace@gmail.com>
Cc: Sourav Poddar <sourav.poddar@ti.com>
Acked-by: Marek Vasut <marex@denx.de>
(cherry picked from commit 99ed1a167578f85963a0cdf5fd7b2291eaecc400)
Change-Id: Ic0d5d184ad7cf702a6027271a2a5388b37a1a0ca
Reviewed-on: http://git.am.freescale.net:8181/20044
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Zhengxiong Jin <Jason.Jin@freescale.com>